Extract the planning logic into a separate module
